bmalloc buildfix on 32 bit Linux (x86/ARM) https://bugs.webkit.org/show_bug.cgi?id=141472
Reviewed by Gyuyoung Kim.

* bmalloc/Algorithm.h:
(bmalloc::roundUpToMultipleOf):
* bmalloc/FixedVector.h:
(bmalloc::FixedVector::clear):
* bmalloc/Sizes.h:
(bmalloc::Sizes::sizeClass):


Canonical link: https://commits.webkit.org/159506@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@179927 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/bmalloc/bmalloc/FixedVector.h
@@ -59,7 +59,7 @@ public:
     void shrink(T*);
     void shrink(size_t);
 
-    void clear() { shrink(0ul); }
+    void clear() { shrink(static_cast<size_t>(0)); }
     bool isEmpty() { return !m_size; }
 
 private:
